% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format_df.R
\name{check_format_df}
\alias{check_format_df}
\title{Function to check a data frame.}
\usage{
check_format_df(data, digits = 6)
}
\arguments{
\item{data}{Dataframe containing the dataset to check}

\item{digits}{integer, number of digits we will round the numerical variables. Default is 6}
}
\value{
string "Check completed".
}
\description{
What is checked:
\itemize{
\item The format of the table: data.table, data.frame or tibble?
\item The names of the columns needed: only letters, integers or "_"
\item The format of the columns: contains only factors and numeric columns
\item Missing data per column
\item Other strange values: "NaN", "NA", or +/- Inf
}
}
\details{
And a summary is created on all numeric variables
}
\examples{
check_format_df(iris)
}
