% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df_rd.R
\name{create_df_rd}
\alias{create_df_rd}
\title{Creation of a random dataset}
\usage{
create_df_rd(
  size = 10,
  seed = NULL,
  numerics = NULL,
  booleans = NULL,
  categories = NULL
)
}
\arguments{
\item{size}{Integer, number of profiles to create.}

\item{seed}{Numeric. Seed for reproducibility. If NULL, a seed is randomly created. The default is NA}

\item{numerics}{List. Normally distributed numeric columns. ID are the names and values are either NULL or a vector of length 2 with mean and std.}

\item{booleans}{List. Binomial distributed columns. Keys are the names, values are either NULL or a float which is the probability of True values.}

\item{categories}{List. Multinomial distributed columns. Keys are the names, values are either NULL or a vector of categorical values. The probabilities will be 1/N if a column has N unique values}
}
\value{
data.frame that can be tested by check_format_df
}
\description{
A function to create a dataset containing :
\itemize{
\item a ID column containing ID from 1 to the size
\item normally distributed variables (if NULL, no column)
\item factors variables with the levels given by list input (if NULL, no column)
\item Booleans with binomial distributions with list input (if NULL, no column)
}
}
\examples{
create_df_rd(10,NULL, list("num1" = NULL,"num2" = c(0,1)), list("bool1" = 0.5),list(a = c(1,2,3),b = c("F","M")))
}
