% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_series.R
\name{find_series}
\alias{find_series}
\title{find_series}
\usage{
find_series(name_file, type_file)
}
\arguments{
\item{name_file}{String, representing the name of the file. If file not at the working directory, add the path to it.}

\item{type_file}{String, representing the extension of the file}
}
\value{
String, with the last name of the series
}
\description{
Function to find the latest file in a series. Types recognized are:
}
\examples{
find_series("test", "csv")
}
