% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_correct_pred.R
\name{glm_correct_pred}
\alias{glm_correct_pred}
\title{glm_correct_pred}
\usage{
glm_correct_pred(predictions, original_avg, new_avg)
}
\arguments{
\item{predictions}{numerical vector, glm predictions}

\item{original_avg}{double, the original average prediction}

\item{new_avg}{double, the new average prediction}
}
\value{
numerical vector, new corrected glm predictions
}
\description{
Correct the predictions of a glm after an over-sampling step.
}
\details{
Formula is based on:
www.data-mining-blog.com/tips-and-tutorials/overrepresentation-oversampling/
}
