% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_strip.R
\name{glm_strip}
\alias{glm_strip}
\title{glm_strip}
\usage{
glm_strip(glm_model)
}
\arguments{
\item{glm_model}{GLM model from stats package.}
}
\value{
glm model
}
\description{
GLM models keep track of the training data and many other details.
This function cuts down all "un-necessary" details from the GLM.
}
