% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbutilities_R.R
\docType{package}
\name{jbutilities-package}
\alias{jbutilities}
\alias{jbutilities-package}
\title{jbutilities: Useful functions ammassed through the years}
\description{
This package contains multiple functions I found, or created, through the years and may be useful to keep.
}
\author{
\strong{Maintainer}: Julie Ballard \email{bronnimannj@gmail.com}

}
\keyword{internal}
