% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbutilities_R.R
\docType{package}
\name{jbutilities}
\alias{jbutilities}
\title{Utilities package}
\description{
This package contains I created over time that I thought would be a good idea to keep
}
\section{Functions in the package}{


\describe{
\item{create_df_rd}{Creates a random table}
\item{check_format_df}{Verifies the format of a table}
\item{find_series}{find the latest file in a series}
\item{name_series}{create the next file in a series}
\item{step_time}{print a step and time}
\item{my_ggtheme}{favorite ggplot theme to any ggplot plot}
\item{cut_fat}{fast version of cut}
\item{class_tree_interaction_finder}{tree interactions for binary classification}
}

