% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggtheme.R
\name{my_ggtheme}
\alias{my_ggtheme}
\title{my_ggtheme}
\usage{
my_ggtheme(x_percent = FALSE, y_percent = FALSE)
}
\arguments{
\item{x_percent}{boolean, if TRUE we label the x axis into percents}

\item{y_percent}{boolean, if TRUE we label the y axis with percents}
}
\value{
a ggplot with a new theme
}
\description{
Function to put my favorite ggplot theme to any ggplot plot
}
\examples{
ggplot2::ggplot(data.frame(a = c(1,2),b = c(3,4)), ggplot2::aes(a,b)) + my_ggtheme()
}
