% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_series.R
\name{name_series}
\alias{name_series}
\title{name_series}
\usage{
name_series(Name, Type)
}
\arguments{
\item{Name}{String, representing the name of the file}

\item{Type}{String, representing the extension of the file}
}
\value{
String, with the next name of the series
}
\description{
Function to create files without over writing them.
}
\details{
It will add an underscore and the next number of the series.
}
\examples{
name_series("test", "csv")
}
