% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pavg}
\alias{pavg}
\title{Mean per row}
\usage{
pavg(df, na_rm)
}
\arguments{
\item{df}{DataFrame. Contains only numerical values}

\item{na_rm, }{boolean. Accept of not the NAs}
}
\value{
NumericVector or mean values.
}
\description{
Returns the average of the numeric values, row per row (similar as pmin)
}
\examples{
pavg( data.frame(a = c(1,2,3), b = c(2,3,3)), na_rm = TRUE)
}
