% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rankdf}
\alias{rankdf}
\title{Rank A numerical dataset row by row}
\usage{
rankdf(df)
}
\arguments{
\item{df}{DataFrame, Data frame numerical values (NA included).}
}
\description{
Returns a data frame where each row is ordered. The columns names are "TOP" + the number.
}
\examples{
rankdf(data.frame(x = c(3,4), y = c(5,6)))
}
