% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ranker}
\alias{ranker}
\title{Rank numerical values}
\usage{
ranker(values, df, method)
}
\arguments{
\item{values}{NumericVector, containing all the values to rank}

\item{df}{DataFrame, Data frame of values against we will rank (NA included).}

\item{method}{String, either "min", "avg" or "max". when there are equalities in the values.}
}
\description{
Returns the rank of the a numerical column from a dataset of Numerical values
}
\examples{
ranker(c(1,2),data.frame(x = c(3,4),y = c(5,6)), "min")
ranker(10,data.frame(x = 10 , y = 10, z = 11, w = 12, u = 13),"min")
}
