% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df_rd.R
\name{create_df_rd}
\alias{create_df_rd}
\title{Creation of a random dataset}
\usage{
create_df_rd(
  size = 10,
  seed = as.numeric(NA),
  Numerics = NULL,
  Booleans = NULL,
  Categories = NULL
)
}
\arguments{
\item{size}{Integer, number of rows of the final dataset.}

\item{seed}{Numeric. If NA, a seed is randomly creacted. Else, takes the seed given.}

\item{Numerics}{Character vector of names of normally distributed numeric vectors}

\item{Booleans}{Named vector of probabilities between 0 and 1}

\item{Categories}{List of factor variables, each element containing the levels as character vector}
}
\value{
data.frame that can be tested by check_format_df
}
\description{
A function to create a dataset containing :
\itemize{
\item a ID column containing ID from 1 to the size
\item normally distributed variables (if NULL, no column)
\item factors variables with the levels given by list input (if NULL, no column)
\item Booleans with binomial distributions with list input (if NULL, no column)
}
}
\examples{
create_df_rd(10,NA, c("num1","num2"), c(0.2,0.3),list(a = c(1,2,3),b = c("F","M")))
}
