% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_series.R
\name{find_series}
\alias{find_series}
\title{find_series}
\usage{
find_series(Name, Type)
}
\arguments{
\item{Name}{String, representing the name of the file}

\item{Type}{String, representing the extension of the file}
}
\value{
String, with the last name of the series
}
\description{
Function to find the latest file in a series. Types regognized are:
}
\examples{
find_series("test", "csv")
}
