% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jbutilities.R
\docType{package}
\name{jbutilities}
\alias{jbutilities}
\title{Utilities package}
\description{
This package contains I created over time that I thought would be a good idea to keep
}
\section{Functions in the package}{


\describe{
\item{check_format_df}{To check the format of a data frame}
\item{create_df_rd}{To create a random table}
}
#'
}

