% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_time.R
\name{step_time}
\alias{step_time}
\title{step_time}
\usage{
step_time(step = 0)
}
\arguments{
\item{step}{String, representing the step. Can be a number too}
}
\value{
String, with step number and the date + time
}
\description{
Function to print a step and time
}
\examples{
step_time(1.1)
}
